+++ R/input-validation.R
@@ -22,16 +22,25 @@ input_validation_anticlustering <- function(x, K, objective, method,
 
   # Allow that K is an initial assignment of elements to clusters
   if (length(K) == 1) {
-    validate_input(K, "K", objmode = "numeric", len = 1, greater_than = 1, must_be_integer = TRUE, not_na = TRUE)
+    validate_input(K, "K", objmode = "numeric", len = 1, 
+                   greater_than = 1, must_be_integer = TRUE, not_na = TRUE)
   } else {
     validate_input(K, "K", objmode = "numeric", len = N, not_na = TRUE)
     if (method != "exchange") {
       stop("Passing an initial cluster assignment via the argument `K` ",
            "only works with method = 'exchange'")
     }
+    if (preclustering == TRUE) {
+      stop("Cannot use preclustering because elements have already been assigned ",
+           "to groups via argument `K`.")
+    }
+    if (method == "ilp") {
+      stop("The argument `K` should indicate the number of groups when using `method = 'ilp'`.")
+    }
     if (argument_exists(categories)) {
       if (length(categories) != length(K)) {
-        stop("Length of arguments `categories` and `K` differ, but should be of same length (K can also be of length 1)")
+        stop("Length of arguments `categories` and `K` differ, but ",
+             "should be of same length (K can also be of length 1)")
       }
     }
   }
@@ -55,7 +64,8 @@ input_validation_anticlustering <- function(x, K, objective, method,
   }
 
   if (!inherits(objective, "function")) {
-    validate_input(objective, "objective", input_set = c("distance", "variance"), len = 1, not_na = TRUE)
+    validate_input(objective, "objective", input_set = c("distance", "variance"), 
+                   len = 1, not_na = TRUE)
     if (objective == "variance" && method == "ilp") {
       stop("You cannot use integer linear programming method to maximize the variance criterion. ",
            "Use objective = 'distance', or method = 'exchange' instead")