+++ NEWS.md
@@ -10,6 +10,11 @@
 of the manuscript "Using Anticlustering to Partition Data Sets Into 
 Equivalent Parts" (https://doi.org/10.1037/met0000301) describing the
 algorithms and criteria used in the package `anticlust` 
+- `anticlustering()` can no longer use an argument `K` of length > 1 with 
+`preclustering = TRUE` because this resulted in undocumented behaviour 
+(this is because it does not make sense to specify an assignment of elements 
+to groups via `K` and at the same time require that preclustering 
+assigns elements to groups)
 
 ## Internal changes